End of section 5

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -42,12 +43,13 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
     {
         var auction = mapper.Map<Auction>(createAuctionDto);
-        // TODO: Add the current user to the auction as Seller
-        auction.Seller = "test";
+
+        auction.Seller = User.Identity?.Name ?? "Unknown user";
 
         context.Auctions.Add(auction);
 
@@ -56,12 +58,14 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
         await publishEndpoint.Publish(mapper.Map<AuctionCreated>(newAuction));
 
         var result = await context.SaveChangesAsync();
+
         if (result == 0) return BadRequest("Could not save changes to DB");
 
         return CreatedAtAction(nameof(GetAuctionById),
             new { Id = auction.Id }, mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -70,7 +74,7 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -81,11 +85,13 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
         await publishEndpoint.Publish(mapper.Map<AuctionUpdated>(auction));
 
         var result = await context.SaveChangesAsync() > 0;
+
         if (result) return Ok();
 
         return BadRequest("Problem saving changes");
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -93,13 +99,14 @@ public class AuctionsController(AuctionDbContext context, IMapper mapper,
 
         if (auction == null) return NotFound();
 
-        // TODO: check seller == username
+        if (auction.Seller != User.Identity?.Name) return Forbid();
 
         context.Auctions.Remove(auction);
 
         await publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
         var result = await context.SaveChangesAsync() > 0;
+
         if (!result) return BadRequest("Could not update DB");
 
         return Ok();
